+++ webrtc/WebRtcTransport.cpp
@@ -36,30 +36,6 @@ using namespace std;
 
 namespace mediakit {
 
-// RTC配置项目
-namespace Rtc {
-#define RTC_FIELD "rtc."
-// rtp和rtcp接受超时时间
-const string kTimeOutSec = RTC_FIELD "timeoutSec";
-// 服务器外网ip
-const string kExternIP = RTC_FIELD "externIP";
-// 设置remb比特率，非0时关闭twcc并开启remb。该设置在rtc推流时有效，可以控制推流画质
-const string kRembBitRate = RTC_FIELD "rembBitRate";
-// webrtc单端口udp服务器
-const string kPort = RTC_FIELD "port";
-
-const string kTcpPort = RTC_FIELD "tcpPort";
-
-static onceToken token([]() {
-    mINI::Instance()[kTimeOutSec] = 15;
-    mINI::Instance()[kExternIP] = "";
-    mINI::Instance()[kRembBitRate] = 0;
-    mINI::Instance()[kPort] = 0;
-    mINI::Instance()[kTcpPort] = 0;
-});
-
-} // namespace RTC
-
 static atomic<uint64_t> s_key { 0 };
 
 static void translateIPFromEnv(std::vector<std::string> &v) {
